+++ action_server/docs/CHANGELOG.md
@@ -9,6 +9,7 @@
     - The tool name is `<action-package-name>/<action-name>`.
   - Both `/mcp` and `/sse` endpoints are available.
     - So, if the Action Server is started at `http://localhost:8000`, the MCP endpoints are available at `http://localhost:8000/mcp` and `http://localhost:8000/sse`.
+  - If an `--api-key` is specified, users must a `Bearer <api-key>` Authorization header to access with it.
   - Note: when the `--reload` flag is used, the MCP tools are also properly reloaded.
   - Note: support for tool cancellation with MCP is not there yet.